Refactor
- Move some widgets around
- Cleanup directory structure
