+++ lib/widget/company/company_list.dart
@@ -8,7 +8,7 @@ import "package:inventree/inventree/model.dart";
 
 import "package:inventree/widget/paginator.dart";
 import "package:inventree/widget/refreshable_state.dart";
-import "package:inventree/widget/company_detail.dart";
+import 'package:inventree/widget/company/company_detail.dart';
 
 
 /*
