+++ lib/widget/drawer.dart
@@ -8,10 +8,10 @@ import "package:inventree/inventree/purchase_order.dart";
 import "package:inventree/inventree/stock.dart";
 import "package:inventree/l10.dart";
 import "package:inventree/settings/settings.dart";
-import "package:inventree/widget/category_display.dart";
+import 'package:inventree/widget/part/category_display.dart';
 import "package:inventree/widget/notifications.dart";
-import "package:inventree/widget/purchase_order_list.dart";
-import "package:inventree/widget/location_display.dart";
+import 'package:inventree/widget/order/purchase_order_list.dart';
+import 'package:inventree/widget/stock/location_display.dart';
 
 
 /*
