+++ lib/widget/home.dart
@@ -11,15 +11,15 @@ import "package:inventree/l10.dart";
 import "package:inventree/settings/select_server.dart";
 import "package:inventree/user_profile.dart";
 
-import "package:inventree/widget/category_display.dart";
+import 'package:inventree/widget/part/category_display.dart';
 import "package:inventree/widget/drawer.dart";
-import "package:inventree/widget/location_display.dart";
-import "package:inventree/widget/part_list.dart";
-import "package:inventree/widget/purchase_order_list.dart";
+import 'package:inventree/widget/stock/location_display.dart';
+import 'package:inventree/widget/part/part_list.dart';
+import 'package:inventree/widget/order/purchase_order_list.dart';
 import "package:inventree/widget/refreshable_state.dart";
 import "package:inventree/widget/snacks.dart";
 import "package:inventree/widget/spinner.dart";
-import "package:inventree/widget/company_list.dart";
+import 'package:inventree/widget/company/company_list.dart';
 
 
 class InvenTreeHomePage extends StatefulWidget {
