+++ lib/widget/order/po_line_detail.dart
@@ -7,14 +7,14 @@ import "package:inventree/app_colors.dart";
 import "package:inventree/helpers.dart";
 import "package:inventree/l10.dart";
 import "package:inventree/widget/progress.dart";
-import "package:inventree/widget/part_detail.dart";
+import 'package:inventree/widget/part/part_detail.dart';
 
 import "package:inventree/widget/refreshable_state.dart";
 import "package:inventree/inventree/company.dart";
 import "package:inventree/inventree/part.dart";
 import "package:inventree/inventree/purchase_order.dart";
 import "package:inventree/widget/snacks.dart";
-import "package:inventree/widget/supplier_part_detail.dart";
+import 'package:inventree/widget/company/supplier_part_detail.dart';
 
 /*
  * Widget for displaying detail view of a purchase order line item
