+++ lib/widget/order/purchase_order_detail.dart
@@ -2,7 +2,7 @@ import "package:flutter/material.dart";
 import "package:flutter_speed_dial/flutter_speed_dial.dart";
 import "package:font_awesome_flutter/font_awesome_flutter.dart";
 import "package:inventree/widget/dialogs.dart";
-import "package:inventree/widget/po_line_list.dart";
+import 'package:inventree/widget/order/po_line_list.dart';
 
 import "package:inventree/api.dart";
 import "package:inventree/app_colors.dart";
@@ -13,11 +13,11 @@ import "package:inventree/l10.dart";
 import "package:inventree/inventree/company.dart";
 import "package:inventree/inventree/purchase_order.dart";
 import "package:inventree/widget/attachment_widget.dart";
-import "package:inventree/widget/company_detail.dart";
+import 'package:inventree/widget/company/company_detail.dart';
 import "package:inventree/widget/notes_widget.dart";
 import "package:inventree/widget/refreshable_state.dart";
 import "package:inventree/widget/snacks.dart";
-import "package:inventree/widget/stock_list.dart";
+import 'package:inventree/widget/stock/stock_list.dart';
 
 
 class PurchaseOrderDetailWidget extends StatefulWidget {
