+++ lib/widget/order/purchase_order_list.dart
@@ -5,7 +5,7 @@ import "package:font_awesome_flutter/font_awesome_flutter.dart";
 import "package:inventree/inventree/company.dart";
 import "package:inventree/inventree/model.dart";
 import "package:inventree/widget/paginator.dart";
-import "package:inventree/widget/purchase_order_detail.dart";
+import 'package:inventree/widget/order/purchase_order_detail.dart';
 import "package:inventree/widget/refreshable_state.dart";
 import "package:inventree/l10.dart";
 import "package:inventree/api.dart";
