+++ lib/widget/part/part_suppliers.dart
@@ -7,7 +7,7 @@ import "package:inventree/api.dart";
 import "package:flutter/material.dart";
 import "package:inventree/inventree/part.dart";
 import "package:inventree/inventree/company.dart";
-import "package:inventree/widget/company_detail.dart";
+import 'package:inventree/widget/company/company_detail.dart';
 import "package:inventree/widget/refreshable_state.dart";
 
 class PartSupplierWidget extends StatefulWidget {
