+++ lib/widget/stock/location_display.dart
@@ -10,12 +10,12 @@ import "package:inventree/l10.dart";
 import "package:inventree/inventree/stock.dart";
 import "package:inventree/preferences.dart";
 
-import "package:inventree/widget/location_list.dart";
+import 'package:inventree/widget/stock/location_list.dart';
 import "package:inventree/widget/progress.dart";
 import "package:inventree/widget/refreshable_state.dart";
 import "package:inventree/widget/snacks.dart";
-import "package:inventree/widget/stock_detail.dart";
-import "package:inventree/widget/stock_list.dart";
+import 'package:inventree/widget/stock/stock_detail.dart';
+import 'package:inventree/widget/stock/stock_list.dart';
 import "package:inventree/labels.dart";
