+++ lib/widget/stock/location_list.dart
@@ -2,7 +2,7 @@ import "package:flutter/material.dart";
 
 import "package:inventree/inventree/model.dart";
 import "package:inventree/inventree/stock.dart";
-import "package:inventree/widget/location_display.dart";
+import 'package:inventree/widget/stock/location_display.dart';
 import "package:inventree/widget/paginator.dart";
 
 import "package:inventree/widget/refreshable_state.dart";
