+++ lib/widget/stock/stock_item_history.dart
